+++ client/examples/file_embedding.py
@@ -1,7 +1,7 @@
 from dotenv import load_dotenv
 from openai import OpenAI
 
-from streaming_assistants import patch
+from astra_assistants import patch
 
 load_dotenv("./.env")
