+++ client/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "astra-assistants"
-version = "2.0.0"
+version = "2.0.2"
 description = "Astra Assistants API - drop in replacement for OpenAI Assistants, powered by AstraDB"
 authors = ["phact <estevezsebastian@gmail.com>"]
 readme = "README.md"
@@ -9,9 +9,8 @@ readme = "README.md"
 python = "^3.10"
 httpx = "^0.26.0"
 openai = "^1.20.0"
-litellm = "^1.39.2"
+litellm = "^1.36.0"
 boto3 = "^1.34.31"
-google-generativeai = "^0.4.1"
 aiohttp = "^3.9.4"
