+++ examples/python/streaming_retrieval/basic.py
@@ -1,6 +1,6 @@
 from openai import OpenAI
 from dotenv import load_dotenv
-from streaming_assistants import patch
+from astra_assistants import patch
 
 
 load_dotenv("./.env")
