+++ pyproject.toml
@@ -55,8 +55,8 @@ litellm = "1.36.2"
 boto3 = "^1.29.6"
 prometheus-fastapi-instrumentator = "^6.1.0"
 google-cloud-aiplatform = "^1.38.0"
-google-generativeai = "^0.4.1"
-streaming-assistants = "0.18.7-pre"
+google-generativeai = "^0.5.0"
+astra-assistants = "2.0.2"
 annotated-types = "^0.6.0"
 pydantic-core = "^2.16.3"
 pydantic = "^2.7.1"
