+++ tests/astra-assistants/test_auto_tool_call_v2.py
