+++ tests/astra-assistants/test_chat_completion.py
