+++ tests/astra-assistants/test_embedding.py
