+++ tests/astra-assistants/test_function_calling_v2.py
