+++ tests/astra-assistants/test_streaming_run_function_calling_v2.py
