+++ tests/astra-assistants/vertex.py
