+++ tests/streaming-assistants/test_streaming_run_function_calling_v1.py
@@ -8,45 +8,45 @@ from typing_extensions import override
 
 logger = logging.getLogger(__name__)
 
-def test_function_calling_gpt_4o(patched_openai_client):
+def test_function_calling_gpt_4o(streaming_assistants_openai_client):
     model="gpt-4o"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
-def test_function_calling_gpt_3_5(patched_openai_client):
+def test_function_calling_gpt_3_5(streaming_assistants_openai_client):
     model="gpt-3.5-turbo"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
-def test_function_calling_cohere(patched_openai_client):
+def test_function_calling_cohere(streaming_assistants_openai_client):
     model="cohere_chat/command-r"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
-def test_function_calling_pplx_mix(patched_openai_client):
+def test_function_calling_pplx_mix(streaming_assistants_openai_client):
     model="perplexity/mixtral-8x7b-instruct"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
 @pytest.mark.skip(reason="pplx_online just looks up the weather and doesn't do the function call")
-def test_function_calling_pplx_online(patched_openai_client):
+def test_function_calling_pplx_online(streaming_assistants_openai_client):
     model="perplexity/pplx-70b-online"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
 @pytest.mark.skip(reason="claude does not consistently work with function calling, skip")
-def test_function_calling_claude(patched_openai_client):
+def test_function_calling_claude(streaming_assistants_openai_client):
     model="claude-3-haiku-20240307"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
 @pytest.mark.skip(reason="litellm does not use the latest gemini tool support yet and gemini refuses without it, skip")
-def test_function_calling_gemini(patched_openai_client):
+def test_function_calling_gemini(streaming_assistants_openai_client):
     model="gemini/gemini-pro"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
-def test_function_calling_groq_llama3(patched_openai_client):
+def test_function_calling_groq_llama3(streaming_assistants_openai_client):
     model="groq/llama3-8b-8192"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
 @pytest.mark.skip(reason="llama does not consistently work with function calling, skip")
-def test_function_calling_llama(patched_openai_client):
+def test_function_calling_llama(streaming_assistants_openai_client):
     model = "meta.llama2-13b-chat-v1"
-    function_calling(model, patched_openai_client)
+    function_calling(model, streaming_assistants_openai_client)
 
 
 def function_calling(model, client):
